% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geessbin.R
\name{geessbin}
\alias{geessbin}
\title{Modified Generalized Estimating Equations for Binary Outcome}
\usage{
geessbin(
  formula,
  data = parent.frame(),
  id = NULL,
  corstr = "independence",
  repeated = NULL,
  beta.method = "PGEE",
  SE.method = "MB",
  b = NULL,
  maxitr = 50,
  tol = 1e-05,
  scale.fix = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{formula}{Object of class formula: symbolic description of model to be
fitted (see documentation of \code{lm} and
\code{formula} for details).}

\item{data}{Data frame.}

\item{id}{Vector that identifies the subjects or clusters (NULL by default).}

\item{corstr}{Working correlation structure. The following are permitted:
"\code{independence}", "\code{exchangeable}", "\code{ar1}", and
"\code{unstructured}" ("\code{independence}" by default).}

\item{repeated}{Vector that identifies repeatedly measured variable within
each subject or cluster. If \code{repeated = NULL}, as is the case in
function \code{gee}, data are assumed to be sorted so that
observations on a cluster are contiguous rows for all entities
in the formula.}

\item{beta.method}{Method for estimating regression parameters.
The following are permitted: "\code{GEE}", "\code{PGEE}", and
"\code{BCGEE}" ("\code{PGEE}" by default).}

\item{SE.method}{Method for estimating standard errors. The following are
permitted: "\code{SA}", "\code{MK}", "\code{KC}", "\code{MD}",
"\code{FG}", "\code{PA}", "\code{GS}", "\code{MB}", "\code{WL}",
"\code{WB}", "\code{FW}", and "\code{FZ}" ("\code{MB}" by default).}

\item{b}{Numeric vector specifying initial values of regression coefficients.
If \code{b = NULL} (i.e., default value), the initial values are
calculated using the ordinary or Firth logistic regression assuming
that all the observations are independent.}

\item{maxitr}{Maximum number of iterations (50 by default).}

\item{tol}{Tolerance used in fitting algorithm (\code{1e-5} by default).}

\item{scale.fix}{Logical variable; if \code{TRUE}, the scale parameter is
fixed at 1 (\code{FALSE} by default).}

\item{conf.level}{Numeric value of confidence level for confidence intervals
(0.95 by default).}
}
\value{
an object of class "\code{geessbin}" representing the results of
modified generalized estimating equations with bias-adjusted covariance
estimators. Generic function \code{summary} provides details of the results.
}
\description{
\code{geessbin} provides results of modified generalized estimating equations
with bias-adjusted covariance estimators for longitudinal or clustered
data with binary outcomes
}
\examples{
data(wheeze)

# analysis of PGEE method with Morel et al. covariance estimator
res <- geessbin(formula = Wheeze ~ City + Time, id = ID,
                repeated = Time, corstr = "ar1", data = wheeze,
                beta.method = "PGEE", SE.method = "MB")

# hypothesis tests for regression coefficients
summary(res)

}
\references{
Liang, K. and Zeger, S. (1986). Longitudinal data analysis using
        generalized linear models.
        \emph{Biometrika}, 73, 13-22.
}
