% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtmat.R
\name{sqrtmat}
\alias{sqrtmat}
\title{Square root of nonsymmetric matrix}
\usage{
sqrtmat(M)
}
\arguments{
\item{M}{Square matrix whose square root is to be computed.}
}
\description{
\code{sqrtmat} is used to calculate the square root of \eqn{E_i - H_{ii}},
which is an adjustment factor in Kauermann and Carroll-type method.
}
